modified headers

+++ include/oneapi/mkl/rng/device/detail/mcg31m1_impl.hpp
@@ -27,7 +27,7 @@ class mcg31m1;
 
 namespace detail {
 
-template <std::int32_t VecSize>
+template <std::uint64_t VecSize>
 constexpr sycl::vec<std::uint64_t, VecSize> select_vector_a_mcg31m1() {
     if constexpr (VecSize == 1)
         return sycl::vec<std::uint64_t, 1>(UINT64_C(1));
@@ -56,7 +56,7 @@ constexpr sycl::vec<std::uint64_t, VecSize> select_vector_a_mcg31m1() {
 // hipSYCL (AdaptiveCpp) doesn't support constexpr sycl::vec constructor
 // that's why in case of hipSYCL backend sycl::vec is created as a local variable
 #ifndef __HIPSYCL__
-template <std::int32_t VecSize>
+template <std::uint64_t VecSize>
 struct mcg31m1_vector_a {
     static constexpr sycl::vec<std::uint64_t, VecSize> vector_a =
         select_vector_a_mcg31m1<VecSize>(); // powers of a
